backend/app/integrations/linkedin.py
import os
import requests
import logging
from typing import Dict, List, Optional, Any
from datetime import datetime

from app.schemas.offer import OfferCreate
from app.schemas.tech import TechCreate

logger = logging.getLogger(__name__)

class LinkedInClient:
    """
    Client pour l'API LinkedIn Jobs
    Note: LinkedIn n'offre pas d'API publique pour les offres d'emploi.
    Cette implémentation est un exemple et nécessiterait une approche différente en production.
    """
    
    def __init__(self):
        self.api_key = os.getenv("LINKEDIN_API_KEY")
        
        if not self.api_key:
            logger.warning("L'API key LinkedIn n'est pas configurée")
    
    def search_offers(
        self, 
        keywords: Optional[str] = None,
        location: Optional[str] = None,
        page: int = 0,
        per_page: int = 20
    ) -> List[Dict[str, Any]]:
        """
        Recherche des offres d'emploi sur LinkedIn
        Note: Implémentation fictive - LinkedIn n'a pas d'API publique pour les offres
        """
        logger.info("Recherche d'offres LinkedIn (simulation)")
        
        # Dans une implémentation réelle, vous pourriez utiliser du web scraping
        # ou un partenariat officiel avec LinkedIn
        
        # Retourne des données simulées pour l'exemple
        return []
    
    def map_to_offer_schema(self, linkedin_offer: Dict[str, Any]) -> OfferCreate:
        """
        Convertit une offre LinkedIn au format de notre application
        """
        # Extraction des technologies mentionnées dans l'offre
        tech_list = []
        
        # Création de l'objet OfferCreate
        return OfferCreate(
            title=linkedin_offer.get("title", "Sans titre"),
            company=linkedin_offer.get("company_name", "Entreprise non précisée"),
            location=linkedin_offer.get("location", "Lieu non précisé"),
            description=linkedin_offer.get("description", ""),
            salary_min=linkedin_offer.get("salary_min"),
            salary_max=linkedin_offer.get("salary_max"),
            contract_type=linkedin_offer.get("contract_type", "Non précisé"),
            remote=linkedin_offer.get("remote", False),
            url=linkedin_offer.get("url", ""),
            posted_at=datetime.now(),
            tech_ids=[],
            techs=tech_list
        )
