backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0
pytest==7.3.1
httpx==0.24.1
